Extract DSL wiring and reorganize to support reuse by NiA
