+++ unified-prototype/unified-plugin/build.gradle.kts
@@ -1,3 +1,8 @@
 plugins {
     kotlin("jvm").version(libs.versions.kotlin).apply(false)
 }
+
+subprojects {
+    group = "org.gradle.experimental"
+    version = "0.1"
+}
