+++ unified-prototype/unified-plugin/plugin-android/src/main/java/org/gradle/api/experimental/android/AndroidLibrary.java
@@ -1,41 +1,41 @@
+/*
+ * Copyright 2024 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package org.gradle.api.experimental.android;
 
 import com.android.build.api.dsl.BaseFlavor;
 import com.android.build.api.dsl.CommonExtension;
 import org.gradle.api.Action;
-import org.gradle.api.experimental.common.LibraryDependencies;
-import org.gradle.api.model.ObjectFactory;
 import org.gradle.api.provider.Property;
 import org.gradle.declarative.dsl.model.annotations.Configuring;
 import org.gradle.declarative.dsl.model.annotations.Restricted;
 
-import javax.inject.Inject;
-
-/**
- * The public DSL interface for a declarative Android library.
- */
 @Restricted
-public abstract class AndroidLibrary {
-    private final LibraryDependencies dependencies;
-    private final AndroidTargets targets;
-
-    @Inject
-    public AndroidLibrary(AndroidTargets targets, ObjectFactory objectFactory) {
-        this.targets = targets;
-        this.dependencies = objectFactory.newInstance(LibraryDependencies.class);
-    }
+public interface AndroidLibrary {
+    /**
+     * @see CommonExtension#getCompileSdk()
+     */
+    @Restricted
+    Property<Integer> getCompileSdk();
 
     /**
      * @see CommonExtension#getNamespace()
      */
     @Restricted
     public abstract Property<String> getNamespace();
-
-    /**
-     * @see CommonExtension#getCompileSdk()
-     */
-    @Restricted
-    public abstract Property<Integer> getCompileSdk();
 
     /**
      * @see BaseFlavor#getMinSdk()
@@ -49,27 +49,17 @@ public abstract class AndroidLibrary {
     @Restricted
     public abstract Property<Integer> getJdkVersion();
 
-    /**
-     * Common dependencies for all targets.
-     */
-    public LibraryDependencies getDependencies() {
-        return dependencies;
-    }
+    AndroidLibraryDependencies getDependencies();
 
     @Configuring
-    public void dependencies(Action<? super LibraryDependencies> action) {
+    default void dependencies(Action<? super AndroidLibraryDependencies> action) {
         action.execute(getDependencies());
     }
 
-    /**
-     * Static targets extension block.
-     */
-    public AndroidTargets getTargets() {
-        return targets;
-    }
+    AndroidTargets getTargets();
 
     @Configuring
-    public void targets(Action<? super AndroidTargets> action) {
+    default void targets(Action<? super AndroidTargets> action) {
         action.execute(getTargets());
     }
 }
