+++ unified-prototype/unified-plugin/plugin-android/src/main/java/org/gradle/api/experimental/android/AndroidTarget.java
@@ -4,7 +4,6 @@ import com.android.build.api.dsl.BaseFlavor;
 import org.gradle.api.Action;
 import org.gradle.api.Named;
 import org.gradle.api.NonNullApi;
-import org.gradle.api.experimental.common.LibraryDependencies;
 import org.gradle.api.model.ObjectFactory;
 import org.gradle.api.provider.Property;
 import org.gradle.declarative.dsl.model.annotations.Configuring;
@@ -16,12 +15,12 @@ import javax.inject.Inject;
 @NonNullApi
 public abstract class AndroidTarget implements Named {
     private final String name;
-    private final LibraryDependencies dependencies;
+    private final AndroidLibraryDependencies dependencies;
 
     @Inject
     public AndroidTarget(String name, ObjectFactory objectFactory) {
         this.name = name;
-        this.dependencies = objectFactory.newInstance(LibraryDependencies.class);
+        this.dependencies = objectFactory.newInstance(AndroidLibraryDependencies.class);
     }
 
     /**
@@ -33,12 +32,12 @@ public abstract class AndroidTarget implements Named {
     /**
      * Dependencies for this target.
      */
-    public LibraryDependencies getDependencies() {
+    public AndroidLibraryDependencies getDependencies() {
         return dependencies;
     }
 
     @Configuring
-    public void dependencies(Action<? super LibraryDependencies> action) {
+    public void dependencies(Action<? super AndroidLibraryDependencies> action) {
         action.execute(getDependencies());
     }
